requirements.txt
groq
python-dotenv
phidata
yfinance
duckduckgo-search
fastapi
uvicorn